ghl-podcast-pipeline/requirements.txt
playwright
notebooklm-py
anthropic
requests
python-dotenv
flask
schedule
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
